Create Tooltip and FollowLink components, implement them.
