+++ src/components/follow-link/index.ts
@@ -1,0 +1,3 @@
+import FollowLink from './follow-link';
+
+export default FollowLink;
