+++ src/components/tooltip/index.ts
@@ -1,0 +1,3 @@
+import Tooltip from './tooltip';
+
+export default Tooltip;
