+++ src/page-templates/topic-content-type-page/topic-content-type-page-template.tsx
@@ -1,13 +1,10 @@
-import { useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import {
     Button,
     Checkbox,
-    ESystemIconNames,
     GridLayout,
     HorizontalRule,
-    Link,
     SideNav,
-    SystemIcon,
     TypographyScale,
 } from '@mdb/flora';
 import { NextPage } from 'next';
@@ -49,8 +46,8 @@ import { useRequestContentModal } from '../../contexts/request-content-modal';
 import { Tag } from '../../interfaces/tag';
 import { tagToTopic } from '../../utils/tag-to-topic';
 import { LocationOptions } from '../../hooks/search/types';
-import styles from '../content-page/styles';
 import EventResults from '../../components/event-results';
+import FollowLink from '../../components/follow-link';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const pluralize = require('pluralize');
@@ -256,20 +253,6 @@ const TopicContentTypePageTemplate: NextPage<TopicContentTypePageProps> = ({
         </GridLayout>
     );
 
-    const [showHoverTooltip, setShowHoverTooltip] = useState(false);
-    const [isFollowing, setIsFollowing] = useState(false);
-
-    let tooltipText =
-        'Receive a monthly digest and recommended content based on topics you follow!';
-    if (isFollowing) {
-        tooltipText = 'Unfollow this topic';
-    }
-
-    const onFollowLinkClick = () => {
-        setIsFollowing(!isFollowing);
-        setShowHoverTooltip(false);
-    };
-
     return (
         <>
             <NextSeo
@@ -299,43 +282,9 @@ const TopicContentTypePageTemplate: NextPage<TopicContentTypePageProps> = ({
                                     {topicName}
                                 </TypographyScale>
                             </a>
-                            <Link
-                                onClick={onFollowLinkClick}
-                                onMouseEnter={() => setShowHoverTooltip(true)}
-                                onMouseLeave={() => setShowHoverTooltip(false)}
-                            >
-                                {isFollowing && (
-                                    <SystemIcon
-                                        sx={{
-                                            stroke: 'blue60',
-                                            path: {
-                                                strokeWidth: 3,
-                                            },
-                                        }}
-                                        name={ESystemIconNames.CHECK}
-                                    />
-                                )}
-                                {!isFollowing && (
-                                    <SystemIcon
-                                        sx={{
-                                            fill: 'blue60',
-                                            stroke: 'blue60',
-                                        }}
-                                        name={ESystemIconNames.PLUS}
-                                    />
-                                )}
-                            </Link>
-                            <div sx={styles.tooltip.tooltipPlacement}>
-                                {showHoverTooltip && (
-                                    <div sx={styles.tooltip.tooltipWrapper}>
-                                        <div sx={styles.tooltip.tooltipArrow} />
-                                        <div sx={styles.tooltip.tooltipBody}>
-                                            {tooltipText}
-                                        </div>
-                                    </div>
-                                )}
-                            </div>
+                            <FollowLink topicName={topicName} iconsOnly />
                         </div>
+
                         <SideNav currentUrl="#" items={tertiaryNavItems} />
                     </div>
                     {header}
